rework so NO_EXTENSIONS is considered
